Upgrade to React Router v6
