+++ cloud-iep-client/package.json
@@ -25,7 +25,7 @@
     "prop-types": "^15.6.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "^4.0.3",
     "recharts": "^2.0.9",
     "typescript": "^4.0.3"
