+++ cloud-iep-client/src/routing/Routes.tsx
@@ -1,6 +1,6 @@
 import { useAuth0 } from '@auth0/auth0-react';
 import React from 'react';
-import { Redirect, Route, Switch } from 'react-router-dom';
+import { Navigate, Route, Routes } from 'react-router-dom';
 import GoalDetail, { goalDetailRoute } from '../ui/goal/GoalDetail';
 import Home, { homeRoute } from '../ui/Home';
 import SimpleStudentPage, {
@@ -10,54 +10,55 @@ import StudentDetail, { studentDetailRoute } from '../ui/student/StudentDetail';
 import Auth, { authCallbackRoute } from '../ui/user/Auth';
 import Profile, { profileRoute } from '../ui/user/Profile';
 
-const Routes = () => {
+const RouterRoutes = () => {
   return (
-    <Switch>
-      <Route path={homeRoute} exact={true}>
-        <Home />
-      </Route>
+    <Routes>
+      <Route path={homeRoute} element={<Home />} />
       <Route
         path={profileRoute}
-        render={() => (
+        element={
           <RequireAuth redirectTo={homeRoute}>
             <Profile />
           </RequireAuth>
-        )}
+        }
       />
       <Route
         path={studentsRoute}
-        render={() => (
+        element={
           <RequireAuth redirectTo={homeRoute}>
             <SimpleStudentPage />
           </RequireAuth>
-        )}
+        }
       />
       <Route
         path={authCallbackRoute}
-        render={() => (
+        element={
           <RequireAuth redirectTo={homeRoute}>
             <Auth />
           </RequireAuth>
-        )}
+        }
       />
       <Route
         path={studentDetailRoute}
-        render={() => (
+        element={
           <RequireAuth redirectTo={homeRoute}>
             <StudentDetail />
           </RequireAuth>
-        )}
+        }
       />
       <Route
         path={goalDetailRoute}
-        render={() => (
+        element={
           <RequireAuth redirectTo={homeRoute}>
             <GoalDetail />
           </RequireAuth>
-        )}
+        }
       />
-      <Route render={() => <Redirect key="redirect" to={homeRoute} />} />
-    </Switch>
+      <Route
+        path="*"
+        element={<Navigate key="redirect" to={homeRoute} replace />}
+      />
+    </Routes>
   );
 };
 
@@ -70,7 +71,7 @@ type PrivateRouteProps = {
 // https://gist.github.com/mjackson/d54b40a094277b7afdd6b81f51a0393f#get-started-upgrading-today
 const RequireAuth = ({ redirectTo, children }: PrivateRouteProps) => {
   const { isAuthenticated } = useAuth0();
-  return isAuthenticated ? children : <Redirect to={redirectTo} />;
+  return isAuthenticated ? children : <Navigate to={redirectTo} replace />;
 };
 
-export default Routes;
+export default RouterRoutes;
