+++ cloud-iep-client/src/ui/Body.tsx
@@ -1,6 +1,6 @@
 import { Grid } from '@material-ui/core';
 import React from 'react';
-import Routes from '../routing/Routes';
+import RouterRoutes from '../routing/Routes';
 import ErrorBoundary from './ErrorBoundary';
 
 function Body() {
@@ -9,7 +9,7 @@ function Body() {
       <Grid item xs={false} sm={2} />
       <Grid item xs={12} sm={8}>
         <ErrorBoundary logError={console.log}>
-          <Routes />
+          <RouterRoutes />
         </ErrorBoundary>
       </Grid>
       <Grid item xs={false} sm={2} />
