+++ cloud-iep-client/src/ui/user/Auth.tsx
@@ -1,6 +1,6 @@
 import { CircularProgress, Typography } from '@material-ui/core';
 import React from 'react';
-import { Redirect } from 'react-router-dom';
+import { Navigate } from 'react-router-dom';
 import ApiStatus from '../../types/ApiStatus';
 import useUsersApi from '../../users/useUserApi';
 import { homeRoute } from '../Home';
@@ -11,7 +11,7 @@ const Auth = () => {
   return (
     <>
       {service.status === ApiStatus.Loading && <CircularProgress />}
-      {service.status === ApiStatus.Loaded && <Redirect to={homeRoute} />}
+      {service.status === ApiStatus.Loaded && <Navigate to={homeRoute} />}
       {service.status === ApiStatus.Error && (
         <Typography variant="h6">
           There was an error creating your user profile. {service.error.message}
